+++ libs/common/src/lib/interfaces/portfolio-performance.interface.ts
@@ -1,7 +1,5 @@
 export interface PortfolioPerformance {
   currentGrossPerformance: number;
   currentGrossPerformancePercent: number;
-  currentNetPerformance: number;
-  currentNetPerformancePercent: number;
   currentValue: number;
 }